# FEAT-#3535: Implement partition shuffling mechanism and algebra sort_by