--- a/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition.py
+++ b/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition.py
@@ -326,6 +326,51 @@ class PandasOnRayDataframePartition(PandasDataframePartition):
         if isinstance(self._ip_cache, ObjectIDType):
             self._ip_cache = ray.get(self._ip_cache)
         return self._ip_cache
+    
+    def split(self, split_func, num_splits, *args):
+        """
+        Split the object wrapped by the partition into multiple partitions.
+
+        Parameters
+        ----------
+        split_func : Callable[pandas.DataFrame, List[Any]] -> List[pandas.DataFrame]
+            The function that will split this partition into multiple partitions.
+        num_splits : int
+            The number of resulting partitions (may be empty).
+        *args : List[Any]
+            Arguments to pass to ``split_func``
+
+        Returns
+        -------
+        list
+            A list of partitions.
+        """
+        logger = get_logger()
+        logger.debug(f"ENTER::Partition.split::{self._identity}")
+        @ray.remote(num_returns=num_splits)
+        def _split_df(df, split_func, *args): # pragma: no cover
+            return split_func(df, *args)
+        outputs = _split_df.remote(self._data, split_func, *args)
+        logger.debug(f"SUBMIT::_split_df::{self._identity}")
+        logger.debug(f"EXIT::Partition.split::{self._identity}")
+        return outputs
+    
+    @classmethod
+    def put_splits(cls, splits):
+        """
+        Create a new partition that wraps the input splits after concatenating them.
+
+        Returns
+        -------
+        PandasOnRayDataframePartition
+            New `PandasOnRayDataframePartition` object.
+        """
+        logger = get_logger()
+        logger.debug(f"ENTER::Partition.put_splits")
+        data, length, width, ip = _concat_splits.remote(*splits)
+        logger.debug(f"SUBMIT::_concat_splits")
+        logger.debug(f"EXIT::Partition.put_splits")
+        return cls(data, length, width, ip)
 
 
 @ray.remote(num_returns=2)
@@ -430,4 +475,34 @@ def _apply_list_of_funcs(funcs, partition):  # pragma: no cover
         len(partition) if hasattr(partition, "__len__") else 0,
         len(partition.columns) if hasattr(partition, "columns") else 0,
         get_node_ip_address(),
+    )
+
+@ray.remote(num_returns=4)
+def _concat_splits(*splits):
+    """
+    Concatenate the splits into one dataframe in a worker process.
+
+    Parameters
+    ----------
+    splits : List[pandas.DataFrame]
+        List of dataframes that correspond to splits to concatenate
+    
+    Returns
+    -------
+    pandas.DataFrame
+        The resulting pandas DataFrame.
+    int
+        The number of rows of the resulting pandas DataFrame.
+    int
+        The number of columns of the resulting pandas DataFrame.
+    str
+        The node IP address of the worker process.
+    """
+    import pandas
+    df = pandas.concat(splits)
+    return (
+        df,
+        len(df),
+        len(df.columns),
+        get_node_ip_address()
     )