@YarShev 2022-07-28
> @RehanSD, what is the status here? Are there any perf numbers?

@mvashishtha, if you are going to take over this PR, could you respond to my question?