@mvashishtha 2022-07-28
@YarShev, @RehanSD is going to fix some correctness bugs on the branch, then do performance testing. AFAIK he doesn't know have any perf numbers yet.