@alejandro-ponder 2022-08-15
Found a bug with value counts. See error below.