@RehanSD 2022-09-27
Hi folks! The `sort_by` is nearly complete - we're just tracking down the last few CI bugs, and hope to get this in by end of week!

One open question is how to handle compat mode. When sorting by a column that contains objects, we need to specify a `method` argument to numpy to ensure that numpy picks a whole object, rather than trying to find the mean between the two objects (which may not support addition or division). The problem is that numpy only supports this for versions 1.23 up, so our current sort can't work on object columns in compat mode. Should we write a compat mode for the parts that need it, and pass down compatibility info to the Modin Frame layer, or should we write the compat mode, and then pick between compat and non-compat modes under the hood by checking the python version? We also have some arguments that we pass through to the underlying `sort_values` that we won't be able to pass through, since they don't exist in the compat mode.