@RehanSD 2022-09-27
I did some rough benchmarking, and noticed that the new sort is sensitive to skew in the data (each of the tests below was only done with 1 trial - will do more robust testing after the algo is improved!)

Modin Main Branch with a uniform distribution (5,000,000 x 100, uniform dist [0, 5_000_000)): 40.345454515 seconds
New Sort Branch with a uniform distribution: 13.770275069000036 seconds
New Sort Branch with a Cauchy distribution (same size of data): 53.401249114999985

I took a look at the sizes of the partitions (that we do the sort on) for each of the cases - for the uniform distribution case, we have a fairly even spread, although one partition is too big:
(_concat_splits pid=3537) 0
(_concat_splits pid=3536) 550956
(_concat_splits pid=3538) 547561
(_concat_splits pid=3539) 547355
(_concat_splits pid=3540) 543955
(_concat_splits pid=3534) 552042
(_concat_splits pid=3535) 541358
(_concat_splits pid=3541) 1716773
While for the Cauchy distribution, we have a similar spread, but two partitions are too big:
(_concat_splits pid=4788) 2891
(_concat_splits pid=4789) 0
(_concat_splits pid=4790) 573087
(_concat_splits pid=4783) 581212
(_concat_splits pid=4785) 568910
(_concat_splits pid=4787) 1016174
(_concat_splits pid=4786) 570652
(_concat_splits pid=4784) 1687074

I timed other aspects of the algorithm - the sampling on each partition and the time to compute an overall sample - both of which are nearly identical across both Dataframes. Do folks have any suggestions on how I can skew-proof the sampling stage, or adjust skew in the splitting stage? I'm thinking potentially sampling from minor columns as well could help, although it might be tricky to know when or how to use it - i.e. we really want a sample of the minor key for a specific range of the major key, so we could break up that range into multiple partitions, which is hard to get. Additionally, this still leaves us open to poor performance when sorting only by one column that has a ton of skew. Would appreciate folk's thoughts!