@RehanSD 2022-09-27
These experiments were run on an m5.2xlarge with 512 GiB of memory. The following script was used:
```python
from time import perf_counter
import ray
import numpy as np

from modin.config import IsRayCluster, BenchmarkMode
IsRayCluster.put(True)
BenchmarkMode.put(True)
ray.init(object_store_memory=1.39586e10)

import modin.pandas as pd
# import pandas as pd

# df = pd.DataFrame(np.random.standard_cauchy(size=(5_000_000, 100)), columns=[f"col {i}" for i in range(100)])
df = pd.DataFrame(np.random.uniform(size=(5_000_000, 100), high=5_000_000), columns=[f"col {i}" for i in range(100)])

start = perf_counter()
df.sort_values(by=['col 0', 'col 1', 'col 2'])
end = perf_counter()
print(f"{end - start}")
```