@RehanSD 2022-09-29
One quick update is I noticed that the partition skews were actually due to two things - how we sample, as well as the fact that I was setting the uppermost and lowermost quantiles to inf and -inf respectively - which meant that in practice, the topmost and bottommost partitions were actually two partitions combined. I've fixed that bug, and the time for standard_cauchy with our new sort is 11.446099726000057 seconds, which is on par with performance on the uniform data. Partition skew is still present in this regime; however, the partitions for the standard Cauchy with this new code looks like this:
```python
(_concat_splits pid=6728) 1207
(_concat_splits pid=6727) 573908
(_concat_splits pid=6726) 571120
(_concat_splits pid=6733) 578198
(_concat_splits pid=6731) 566099
(_concat_splits pid=6732) 581331
(_concat_splits pid=6730) 1018288
(_concat_splits pid=6729) 1109845
```
I will add that there seems to be some stochasticity to the performance when measuring this branch - sometimes the timing will be 11 seconds, and sometimes it can jump up to 50, but I'm fairly certain that this is due to ray not tearing down properly, since it always goes to 11 if I just run `ray stop -f` a couple of times, and wait a few minutes before rerunning.

Still working on the sample/split code - will post the benchmarks @mvashishtha asked for once I finish!