@mvashishtha 2022-10-03
@RehanSD it's great to hear that Modin is doing about equally well on the Cauchy data. Since the new Modin performance on Cauchy data is not very different from the performance on uniform data, it seems very unlikely that Modin is going to do much worse than pandas on the Cauchy data.

Since at least for your benchmark, the new implementation's performance on uniform data is about on par with pandas and ~3x better than what we have now, I think we should prioritize making the current implementation readable and merging it. I think we should fine tune the performance later