@RehanSD 2022-11-30
On a 5_000_000 million row x 100 col df, we get the following times:
| pandas | Modin Master | Modin This Branch |
| ------- | -------------- | ------------------ |
| 16.07 ± 0.0747 s | 40.36 ± 0.1191 s | 9.97 ± 0.0821 s |

This is on an M5.2XLarge, with a 32GB /dev/shm, and 30.02GB for Ray's object store.

Modin Bench script:
```python
from time import perf_counter
import ray
import numpy as np

from modin.config import IsRayCluster, BenchmarkMode
IsRayCluster.put(True)
BenchmarkMode.put(True)
ray.init(object_store_memory=2.79586e10)

import modin.pandas as pd
# import pandas as pd

df = pd.DataFrame(np.random.standard_cauchy(size=(5_000_000, 100)), columns=[f”col {i}” for i in range(100)])
# df = pd.DataFrame(np.random.uniform(size=(5_000_000, 100), high=5_000_000), columns=[f”col {i}” for i in range(100)])

start = perf_counter()
df.sort_values(by=[‘col 0’, ‘col 1’, ‘col 2’])
end = perf_counter()
print(f”{end - start}“)
```

pandas Bench script:
```python
from time import perf_counter
import numpy as np

import pandas as pd

df = pd.DataFrame(np.random.standard_cauchy(size=(5_000_000, 100)), columns=[f"col {i}" for i in range(100)])

start = perf_counter()
df.sort_values(by=['col 0', 'col 1', 'col 2'])
end = perf_counter()
print(f"{end - start}")
```